+++ app/services/verifier/acapy_verifier.py
@@ -9,7 +9,7 @@ from app.models.verifier import (
     RejectProofRequest,
     SendProofRequest,
 )
-from shared.models.topics import PresentationExchange
+from shared.models.webhook_topics import PresentationExchange
 
 
 class Verifier(ABC):
