+++ app/tests/e2e/test_webhook.py
@@ -4,7 +4,7 @@ from fastapi import HTTPException
 from app.routes import webhooks
 from app.tests.util.ecosystem_connections import BobAliceConnect
 from shared import RichAsyncClient
-from shared.models.topics import Connection
+from shared.models.webhook_topics import Connection
 
 BASE_PATH = webhooks.router.prefix
