+++ app/tests/issuer/test_issuer.py
@@ -9,7 +9,10 @@ from app.models.issuer import IndyCredential
 from app.services.issuer.acapy_issuer_v1 import IssuerV1
 from app.services.issuer.acapy_issuer_v2 import IssuerV2
 from app.tests.util.mock import to_async
-from shared.models.topics import CredentialExchange, IssueCredentialProtocolVersion
+from shared.models.webhook_topics import (
+    CredentialExchange,
+    IssueCredentialProtocolVersion,
+)
 from shared.util.mock_agent_controller import MockContextManagedController
