+++ webhooks/routers/sse.py
@@ -7,7 +7,7 @@ from sse_starlette.sse import EventSourceResponse
 
 from shared import DISCONNECT_CHECK_PERIOD, QUEUE_POLL_PERIOD, SSE_TIMEOUT, APIRouter
 from shared.log_config import get_logger
-from shared.models.topics import WEBHOOK_TOPIC_ALL
+from shared.models.webhook_topics import WEBHOOK_TOPIC_ALL
 from webhooks.dependencies.container import Container
 from webhooks.dependencies.event_generator_wrapper import EventGeneratorWrapper
 from webhooks.dependencies.sse_manager import SseManager
